requirements.txt
anthropic>=0.39.0
reportlab>=4.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
